finished cbmDisk tasks for now

+++ src/main/java/org/cbm/ant/cbm/CBMDiskTaskCommand.java
@@ -10,6 +10,6 @@ public interface CBMDiskTaskCommand
 
 	public boolean isExecutionNecessary(long lastModified, boolean exists);
 
-	public int execute(CBMDiskTask task, CBMDiskOperator operator) throws BuildException;
+	public Long execute(CBMDiskTask task, CBMDiskOperator operator) throws BuildException;
 
 }
